nina/pyproject.toml
[project]
name = "nina-mcp"
description = "Warnmeldungen des Bevölkerungsschutzes für Gefahrenlagen wie zum Beispiel Gefahrstoffausbreitung oder Unwetter"
version = "0.1.0"

dependencies = [
    "fastmcp>=2.11.3",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]
